Finished implementing / testing TextFile Parser
Cleaning up TextFile Dumper

+++ apptbook/src/main/java/edu/pdx/cs410J/prathik/TextDumper.java
@@ -9,39 +9,47 @@ public class TextDumper implements AppointmentBookDumper<AppointmentBook> {
 
     String textFile = "";
 
-    TextDumper(){
+    TextDumper() {
         super();
     }
 
-    public void SetFile(String fileName){
-        if ((fileName == null) || (fileName.equals("")))
+    public void SetFile(String textFile) {
+        if (!CheckValidFileName(textFile))
             throw new InvalidFileName("Please provide a valid file name");
-        this.textFile = fileName;
+
+        this.textFile = textFile;
     }
 
-    public String GetFile(){
+    public String GetFile() {
         return this.textFile;
     }
+    
+    private boolean CheckValidFileName(String textFile){
+        return !((textFile == null) || (textFile.equals("")));
+    }
 
-    public void dump(AppointmentBook appointmentBook){
+    public void dump(AppointmentBook appointmentBook) {
+        if (!CheckValidFileName(this.textFile))
+            throw new InvalidFileName("Please provide a valid file name using SetFile method");
+        
         File f = new File(this.textFile);
         // If textFiles exists, delete and then recreate it
-        if(f.isFile()) {
+        if (f.isFile()) {
             f.delete();
         }
         try {
             f.createNewFile();
 
 
-        FileWriter fw = new FileWriter(textFile);
+            FileWriter fw = new FileWriter(textFile);
 
-        fw.write(appointmentBook.getOwnerName() + "\n");
-        for(Appointment appointment : appointmentBook.getAppointments())
-            fw.write(appointment.getDescription() + "\t" + appointment.getBeginTimeString() + "\t" + appointment.getEndTimeString() + "\n");
+            fw.write(appointmentBook.getOwnerName() + "\n");
+            for (Appointment appointment : appointmentBook.getAppointments())
+                fw.write(appointment.getDescription() + "\t" + appointment.getBeginTimeString() + "\t" + appointment.getEndTimeString() + "\n");
 
-        fw.close();
-        } catch (Exception e){
-            throw new InvalidFileName("Please provide a valid file name");
+            fw.close();
+        } catch (Exception e) {
+            throw new InvalidFileName("Please provide a valid file name using SetFile method");
         }
     }
 }
